# [Datasets] Prune unused columns before aggregate