--- a/python/ray/data/_internal/arrow_block.py
+++ b/python/ray/data/_internal/arrow_block.py
@@ -269,8 +269,8 @@ class ArrowBlockAccessor(TableBlockAccessor):
     def select(self, keys: List[KeyFn]) -> "pyarrow.Table":
         if not all(isinstance(key, str) for key in keys):
             raise ValueError(
-                "keys must be a list of strings when aggregating on Arrow blocks, "
-                f"but got: {type(keys)}."
+                "keys must be a list of column name strings when aggregating on Arrow "
+                f"blocks, but got: {keys}."
             )
         return self._table.select(keys)