--- a/python/ray/data/block.py
+++ b/python/ray/data/block.py
@@ -276,6 +276,10 @@ class BlockAccessor(Generic[T]):
         """
         raise NotImplementedError
 
+    def select(self, keys: List[KeyFn]) -> Block:
+        """Return a new block containing the provided keys."""
+        raise NotImplementedError
+
     def random_shuffle(self, random_seed: Optional[int]) -> Block:
         """Randomly shuffle this block."""
         raise NotImplementedError