--- a/python/ray/data/grouped_dataset.py
+++ b/python/ray/data/grouped_dataset.py
@@ -5,7 +5,17 @@ from ray.data._internal.compute import CallableClass, ComputeStrategy
 from ray.data._internal.plan import AllToAllStage
 from ray.data._internal.shuffle import ShuffleOp, SimpleShufflePlan
 from ray.data._internal.push_based_shuffle import PushBasedShufflePlan
-from ray.data.aggregate import AggregateFn, Count, Max, Mean, Min, Std, Sum
+from ._internal.table_block import TableBlockAccessor
+from ray.data.aggregate import (
+    _AggregateOnKeyBase,
+    AggregateFn,
+    Count,
+    Max,
+    Mean,
+    Min,
+    Std,
+    Sum,
+)
 from ray.data.block import (
     Block,
     BlockAccessor,
@@ -33,6 +43,9 @@ class _GroupbyOp(ShuffleOp):
     ) -> List[Union[BlockMetadata, Block]]:
         """Partition the block and combine rows with the same key."""
         stats = BlockExecStats.builder()
+
+        block = _GroupbyOp._pruneUnusedColumns(block, key, aggs)
+
         if key is None:
             partitions = [block]
         else:
@@ -58,6 +71,38 @@ class _GroupbyOp(ShuffleOp):
         return BlockAccessor.for_block(mapper_outputs[0]).aggregate_combined_blocks(
             list(mapper_outputs), key, aggs, finalize=not partial_reduce
         )
+
+    @staticmethod
+    def _pruneUnusedColumns(
+        block: Block,
+        key: KeyFn,
+        aggs: Tuple[AggregateFn],
+    ) -> Block:
+        """Prune unused columns from block before aggregate."""
+        prune_columns = True
+        columns = set()
+
+        if isinstance(key, str):
+            columns.add(key)
+        elif callable(key):
+            prune_columns = False
+
+        for agg in aggs:
+            if isinstance(agg, _AggregateOnKeyBase) and isinstance(agg._key_fn, str):
+                columns.add(agg._key_fn)
+            elif not isinstance(agg, Count):
+                # Not prune columns if any aggregate key is not string.
+                prune_columns = False
+
+        block_accessor = BlockAccessor.for_block(block)
+        if (
+            prune_columns
+            and isinstance(block_accessor, TableBlockAccessor)
+            and block_accessor.num_rows() > 0
+        ):
+            return block_accessor.select(list(columns))
+        else:
+            return block
 
 
 class SimpleShuffleGroupbyOp(_GroupbyOp, SimpleShufflePlan):
@@ -249,31 +294,14 @@ class GroupedDataset(Generic[T]):
         else:
             sorted_ds = self._dataset.repartition(1)
 
-        import numpy as np
-
-        # Returns the keys of the batch in numpy array.
-        # Returns None if the self._key is None.
-        def get_keys(batch) -> Optional[np.ndarray]:
-            import pandas as pd
-            import pyarrow as pa
-
-            if self._key is None:
-                return None
-            if isinstance(batch, pd.DataFrame) or isinstance(batch, pa.Table):
-                assert isinstance(self._key, str)
-                return batch[self._key].to_numpy()
-            elif isinstance(batch, List):
-                assert callable(self._key)
-                return np.array([self._key(item) for item in batch])
-            else:
-                raise ValueError(
-                    f"Unsupported batch type for map_groups: {type(batch)}"
-                )
-
         # Returns the group boundaries.
-        def get_key_boundaries(batch):
+        def get_key_boundaries(block_accessor: BlockAccessor):
+            import numpy as np
+
             boundaries = []
-            keys = get_keys(batch)
+            # Get the keys of the batch in numpy array format
+            keys_block = block_accessor.select([self._key])
+            keys = BlockAccessor.for_block(keys_block).to_numpy()
             start = 0
             while start < keys.size:
                 end = start + np.searchsorted(keys[start:], keys[start], side="right")
@@ -286,7 +314,7 @@ class GroupedDataset(Generic[T]):
         def group_fn(batch):
             block_accessor = BlockAccessor.for_block(batch)
             if self._key:
-                boundaries = get_key_boundaries(batch)
+                boundaries = get_key_boundaries(block_accessor)
             else:
                 boundaries = [block_accessor.num_rows()]
             builder = block_accessor.builder()