Signed-off-by: Cheng Su <scnju13@gmail.com>

<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR. If you don't have the access to it, we will shortly find a reviewer and assign them to your PR. -->

## Why are these changes needed?

<!-- Please give a short summary of the change and the problem this solves. -->
This PR is to prune (remove) unused columns before doing aggregate (in `_GroupbyOp.map()`). Only keeps the group-by column and columns used in aggregate functions. All other columns can be pruned, so it reduces the cost during sort and aggregate.

Also introduce `BlockAccessor.select(keys)` to get a new `Block` with only selected `keys`/columns. Refactored existing code path in `map_groups` to also use the API. Later on, we can use this API to implement `Dataset.select_columns`.

Tested with query in h2oai benchmark - https://github.com/ray-project/ray/pull/28486 . Reduced query runtime by 50% with this PR.

After this PR, the query takes 7 seconds:

```
>>> path = "/Users/chengsu/db-benchmark/G1_1e7_1e2_0_0.csv"
>>> input_ds = ray.data.read_csv(path).repartition(10).fully_executed()
>>> input_ds.groupby("id1").sum("v1")
Sort Sample: 100%|███████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 13.08it/s]
Shuffle Map: 100%|███████████████████████████████████████████████████████████████| 10/10 [00:07<00:00,  1.30it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 1022.28it/s]
Dataset(num_blocks=10, num_rows=100, schema={id1: string, sum(v1): int64})
```

Before this PR, the query takes 13 seconds:

```
>>> input_ds.groupby("id1").sum("v1")
Sort Sample: 100%|███████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 11.50it/s]
Shuffle Map: 100%|███████████████████████████████████████████████████████████████| 10/10 [00:13<00:00,  1.31s/it]
Shuffle Reduce: 100%|███████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 851.34it/s]
Dataset(num_blocks=10, num_rows=100, schema={id1: string, sum(v1): int64})
```

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/master/.
- [x] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(