@c21 2022-09-16
Also tested with Polars enabled, speed up 25% of end-to-end runtime for h2oai benchmark:

With this PR:

```
Running case: h2oai-500M-q1
Sort Sample: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 12.65it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 15.56it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 416.53it/s]
Result of case h2oai-500M-q1: {'time': 1.4650313749999988}
Running case: h2oai-500M-q3
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 1472.72it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 10.05it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 293.45it/s]
Result of case h2oai-500M-q3: {'time': 1.0395020000000006}
Running case: h2oai-500M-q4
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 2191.04it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 44.32it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 738.77it/s]
Result of case h2oai-500M-q4: {'time': 0.2474954579999995}
Running case: h2oai-500M-q5
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 2225.69it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 20.95it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 357.01it/s]
Result of case h2oai-500M-q5: {'time': 0.5134781659999987}
Running case: h2oai-500M-q7
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 2572.09it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 11.66it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 273.02it/s]
Map_Batches: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 293.30it/s]
Result of case h2oai-500M-q7: {'time': 0.9361942499999998}
Finish benchmark: aggregate in 4.201701248999997 seconds
```

Without this PR:

```
Running case: h2oai-500M-q1
Sort Sample: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 11.90it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:01<00:00,  8.71it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 640.99it/s]
Result of case h2oai-500M-q1: {'time': 2.014937833000001}
Running case: h2oai-500M-q3
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 1845.11it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:01<00:00,  8.49it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 267.40it/s]
Result of case h2oai-500M-q3: {'time': 1.2240467079999995}
Running case: h2oai-500M-q4
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 1839.93it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 19.83it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 640.98it/s]
Result of case h2oai-500M-q4: {'time': 0.5297700419999991}
Running case: h2oai-500M-q5
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 2446.80it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 15.20it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 333.35it/s]
Result of case h2oai-500M-q5: {'time': 0.6956326670000017}
Running case: h2oai-500M-q7
Sort Sample: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 2377.99it/s]
Shuffle Map: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:01<00:00,  9.84it/s]
Shuffle Reduce: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 278.37it/s]
Map_Batches: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:00<00:00, 266.96it/s]
Result of case h2oai-500M-q7: {'time': 1.0983343329999968}
Finish benchmark: aggregate in 5.562721582999998 seconds
```